Fixing Implicit Conversion Warnings
Fixing implicit conversion warnings to ensure that the program always correctly converts ints to longs. This was caused by changes to ARM and introduction of mandatory 64bit support by Apple for iOS apps.

+++ zbar/config.c
@@ -156,7 +156,7 @@ int zbar_parse_config (const char *cfgstr,
 #ifdef HAVE_ERRNO_H
         errno = 0;
 #endif
-        *val = strtol(eq + 1, NULL, 0);
+        *val = (int)strtol(eq + 1, NULL, 0);
 #ifdef HAVE_ERRNO_H
         if(errno)
             return(1);
